+++ src/app/mypage/page.tsx
@@ -3,7 +3,7 @@ import DuckPointArea from '@/containers/mypage/DuckPointArea'
 import InfoArea from '@/containers/mypage/InfoArea'
 import MannerDuckArea from '@/containers/mypage/MannerDuckArea'
 import TradeArea from '@/containers/mypage/TradeArea'
-import { getMannerDuck, getDuckPoint } from '@/utils/mypageApiActions'
+import { getMannerDuck, getDuckPoint } from '@/utils/memberApiActions'
 
 export default async function MyPage() {
   const [mannerDuckData, duckPointData] = await Promise.all([
